+++ docs/user_guide/changelog.rst
@@ -16,7 +16,7 @@ New features
 API changes
 -----------
 
-- Data loader: default process creation strategy for worker process changed from "spawn" to "forkserver".
+- Data loader: default process creation strategy for worker process changed from "spawn" to "fork".
 
 Bug Fixes
 ---------